[ssl cert] fixes the snap version multipass can not overwrite the key file issue.
This is due to the key file is read-only to the owner itself. As a result of that, the owner can not open it with "wb" write and binary. In our use case, we need to overwrite the file, so remove it first can be a good workaround of this.